BUG CORE-2828: Reduce contrast for dim labels in column chart.
Reviewed-by: chris.laidlaw@socrata.com
QA: Just making the dim labels on the column chart a bit less visible.
The dim labels are shown when you click on a column that does not have a
label (only the first 3 columns have labels). In this case the first 3
labels are dimmed. I made these 3 labels more dimmed in this state.