FEATURE: Currently all GET requests to socrata's servers have an additional parameter, "_=[timestamp]" which is jquery's way of circumventing browser caches (http://api.jquery.com/jquery.ajax/). This setting is applied globally; but it is unnecessary in a few cases:   - requests to the views (ViewService) are fully ETAGged   - requests to the rows (/id, RowsService) are fully ETAGged   - Comments Service is stupid; cache it.   - List of views and publication groups. This is more controversial; we may want to add some explicit Cache-Control on this; but I *think* this is a weak enough use case remove the caching block.
I searched through the code and added "cache: false" explicitly to all other commands.

QA: Please do a pretty through regression of the frontend, looking specfically for cases where data is out of date. Since this a caching improvement it's sort-of risky; and should require some settling time in staging before we move it out to prod.

Reviewed-by: Randy.antler@socrata.com