FEATURE: CORE-3011 Support multiple downtime maintenance notices
The maintenance message logic is designed to accept a singular upcoming maintenance window and display the notification. We normally start showing such maintenance 48 hours in advance of such events.

Azure, the complicated setup they are, has decided to 2 12-hour maintenance windows over the course of 2 days. This means that we need to show two maintenance windows simultaneously, or just pretend it's one 36-hour long window. :-(

This change allows more than one maintenance message to be configured, displayed and individually dismissed.

Reviewed-by: paul.paradise@socrata.com

QA: If you have the capacity to alter downtime notices, please test to see that multiple downtime notices can be made to display at once and can be dismissed individually. If you do not have the capacity to test this, then leave it to th OPS team to test this in staging.