BUG: CORE-3168 Use correct value for 'Last Updated'
* Add http interceptor that observes values for 'X-SODA2-Truth-Last-Modified'
* Add function for taking a string date and returning time since that
  date
* Bind LastModified observer, updating on distinct, and mapping as time
  from now
* Create <last-updated/> directive that observes the LastModified
  service
* Add fromNow filter for filtering a date as a moment "from now" string

Reviewed-by: randy.antler@socrata.com

QA: Check that 'Last Updated' value corresponds with actual last updated value
of the dataset.  A way to find the last updated date for a dataset is to visit
the legacy interface for a dataset '/dataset/FOUR-FOUR', and click on the
'About' button.  A quick way to find the time difference is to use Wolfram
Alpha: (http://www.wolframalpha.com/input/?i=days%20since%20<the date, e.g. 8-17-2014>)