FEATURE: CORE-2506 Visually tweak timeline flyout
* Refactor timelineChart to be in line with Angular styling
* Add 'overflow:"hidden"' to '.card' to set bounds for flyout
* Add computed margin for offsetting flyout from right edge of card
* Change label hiding to be by opacity instead of display:hidden
* Modify onOpen flyout to show transparent labels when the segment is
  moused over, and return to transparent on mouse out

Reviewed-by: jerjou.cheng@socrata.com

QA:
1. Mouse over the rightmost segment on a timeline chart. The flyout
should appear and not show outside the card, instead being 15 pixels
from the right edge of the card.
2. Viewing a timeline chart that alternates label visibility between
each tick. Mousing over a segment that has a hidden label should cause
the label to display and the other labels to dim. Mousing over where the
a hidden label is should also cause the label to appear and the others
to dim.  Mousing off should return it to its original state.