FEATURE: CORE-2467 Render link to source dataset in metadata info section
In the metadata section of a view page, render a link to the source dataset along with the dataset name. Truncate long dataset names with an ellipsis and show a flyout containing full neame when truncation is active.

Incidentally add "show-more-mode=flyout" feature to multilineEllipsis directive.

QA: When viewing pages in the new UX, one should now see a "SOURCE DATASET" link in the upper-right corner that reflects the name of the underlying dataset. Setting the dataset name to a very long value should cause the name to appear truncated with an ellipsis and a flyout on hover that reveals the full dataset name.

Reviewed-by: chris.laidlaw@socrata.com
Signed-off-by: Randy Antler <randy.antler@socrata.com>