FEATURE: CORE-3392 Add GeoJSON to API panel
* Refactor API panel into a stand-alone directive
* Refactor "selection label" in API panel into its own directive
* Add GeoJSON capability to the dataset service
* Add logic to hide or display the 'JSON' and 'GeoJSON' toggle buttons

Reviewed-by: jerjou.cheng@socrata.com

QA: Click on the API button when viewing a card view that has geography
information. There will be both a JSON and GeoJSON data in the API
panel.  Clicking on them will change the URL visible in the panel.
Click on the URL label will still toggle highlighting of the URL.