BUG: Save As failure message should still time out.
QA: Make a change to the page and open the Save As dialog.
Then, disconnect from the network to cause the save to fail.
Then click Save. We expect the button's text to change to "Failed"
and the background to change to red. After about 8 seconds,
the button should revert to its normal appearance.
Reviewed-by: randy.antler@socrata.com