FEATURE: CORE-3365 Add support for Airbrake Javascript error notification
This change adds a feature flag that, when enabled, turns on Airbrake logging of Javascript exceptions in the Dataspace Angular application.

The feature flag is called "enable_airbrake_js".

QA: When the feature flag is disabled, there should be no change in the application behavior. When enabled, you should see notices in the brower console that exceptions are being reported to Airbrake whenever an exception occurs on the page.

The notice should resemble the following: "airbrake: error #1267460919317339239 was reported: https://airbrake.io/locate/1267460919317339239"

Reviewed-by: Giacomo Ferrari <giacomo.ferrari@socrata.com>
Signed-off-by: Randy Antler <randy.antler@socrata.com>