FEATURE: CORE-2455 Add Search card
* Add search card visualization
* Update card mapping to support search card
* Update existing tests to support search card
* Add tests for search card

QA: You should now be able to add a search card to the cards view.  With
a search card present, you should see a search input, magnifying
glass button, and example search terms.  Entering in a search term (exact
match only, currently), and either hitting enter or clicking the magnifying
glass button will execute a search, expanding the card and showing the results.
If there are search results, a table will show.  If no results, a messag will
show asking to try again.  If the datatype for that column is a number,
and you enter text, you will get a message explaining that.  Clicking
the 'x', hitting the excape key, or deleting the entered search text,
will remove the search results (table or message).