Fix DownloadServiceTest - make more deterministic, rather than depending on the non-existence of a certain url on the karma server.
Also, apparently Rx.timer behaves differently if given an initial delay that's equal to the
interval. I suspect that's the cause of the intermittent failures, and have run the tests
continuously for awhile with no failures (before it would fail every 10 runs or so), so hopefully
that stabilizes the test.