FEATURE: CORE-2521 Add aggregation chooser
* Add UI element for aggregation choosing
* Add plumbing in Page model for aggregation functionality
* Update visualizations to support setting an aggregation function and
  field

QA: When viewing the datacards view for a dataset that has at least one
number type card, a dropdown will be available in the quick filter bar
to alternate between choosing the number of (count) of the units to
aggregate data as, and the sum of a selected number column.  Toggling
between these will change the values that display in the column,
timeline, and choropleth cards.