BUG: CORE-3875 Make visualizations not collapse
QA: In a datacard page, on page load, the barchart and timeline chart visualizations (at least)
would sometimes load with the bar charts collapsed vertically, so you could see the X axis but not
the bars/lines. Is probably a race condition, but you should be able to just keep refreshing the
page and verifying that the charts load and are visible.