FEATURE: CORE-4072 Add check and redirect for NBE
* Redirect users to new UX if dataset is NBE and user is not superadmin,
  otherwise if the dataset is NBE and user _is_ superadmin, show warning
* If the dataset does not have a new UX page created yet, redirect to
  the main homepage
* Include tests for the datasets controller that cover the three
  scenarios

QA: When attempting to view a dataset that is intended for the new UX, a
superadmin user should see a warning at the top of the page, and a
non-superadmin user should be redirected to the new UX, if a new UX page
has been created for this dataset.  Otherwise, it redirects to the
standard home page.