FEATURE: 36.01, 36.03 - Header logo and links
* Factor out existing header markup into `<page-header />` directive
* Use configurations service to drive data in pageHeader directive, with
  fallbacks for if the service errors or is unavailable
* Add tests for directive
* Update tests that get the directive by other means
* Add [documentation](/doc/new-ux-customization-howto.md) for setting up
  configuration values

QA: There should now be a Socrata logo at the top of the page, if there
are no header customizations in place.  If the user is logged out, there
should be a 'Sign In' and 'Sign Up' link at the top of the page.  If the
user is signed in, there should be a 'Sign Out' link at the top of the
page.  Going through the sign in or sign up workflows from the cards
view page should return the user to the cards view page.