FEATURE: Front-end updates to support federated SSO through Auth0.
This check-in contains the changes to the Auth0 controller to consume the
new endpoint added to Core.  This code is responsible for cracking the token,
validating it and then proxying the JSON blob back to the authenticate service.

It assumes a fully trusted connection between the front-end and Core.  If this
assumption is not valid, then we will be opening ourselves to a MITM attack, as a
malicious user could tamper the unsigned JSON token extracted from the front-end.
Should this assumption become invalid, then we'll want to implement the
delegation flow supported from Auth0.