Disable leaflet retina detection for feature maps.
There are a few problems with it.  First, it is buggy and actually
_prevents_ you from getting tiles at the most granular level (18) from
the tile server.  Second it loads 4x as many tiles which is a significant
perf hit since these are all dynamically created tiles.

Might be able to re-evaluate in the future and dig into it when we have
more time.