CORE-3636 - Fix invalid dates breaking timeline chart
Establish a maximum valid Javascript date constant and use this constant to filter out invalid dates from queries to the backend.
Additionally, when invalid dates that are less than the maximum value are still encountered, a message is displayed to the user conveying this fact.
toUriRegex helper function updated to support less-than, greater-than and colon symbols.

Signed-off-by: Randy Antler <randy.antler@socrata.com>