Fix incorrect query ranges generated by timeline chart.
This appears to have no effect on how the chart selection is now rendered; this code previously added one 'datasetPrecision' unit to the selection end date to work around a selection rendering issue.

Add a 1 pixel top margin to the chart clear selection label so that it does not overlap the x-axis.