This is the worst Apache config file I have seen in my life.
Ok, I'm totally lying I've seen far worse.  Update Apache config
to proxy to tileserver directly instead of proxying through rails,
as part of the plan to move to ATS proxying to the tileserver instead
of going through rails.

But seriously, so cut and pastey and unnecessarily large.  On the
todo backburner for refactor.  Ignoring the fact we should pick
either nginx or Apache and not both.

nginx file already ... appears to do this same thing.