Ensure that pages are always serialized with a 'version' property.
Default to the 'current version' if no version is present in the existing page metadata.

Add feature flag supporting the 'current version' of dataset and page metadata.

Update tests.

Update Mockumentary to allow creation of metadata in addition to instantiation of objects.

Remove errant `.only` (THERE CAN BE ONLY NONE).