CORE-4945
Make terminology consistent.

Remove periods from the end of flyout texts.

The .erb file was using `.capitalize`, which was making the capitalization of the rest of the string weird. This is now special-cased for Data Lens pages.

Fix test that used the outdated flyout content.

One registerFlyout handler was redundant in `CardsViewController.js`.