ONCALL-2309
On 2015-04-20 we became aware of a certain type of request that caused Phidippides to decide core servers were dead when they were not, with the eventual effect of it deciding that there were no core servers available and therefore responding to legitimate metadata requests with 500 errors.

The cause was determined to be requests for federated datasets that triggered malformed redirects to the 'official' dataset sources that in turn caused circular redirects.

In order to avoid these circular redirects while we are investigating and deploying a more permanent solution, and because the commented-out code below was already up for replacement with simpler logic, we are temporarily forgoing the logic used to determine a redirect destination that depends on a dataset's metadata.

Instead, we will rely only on the presence or absence of a view's migration and its old-back-end four-by-four.