CORE-5452 - Improve docker container checking for /version
Due to `CurrentDomain` permeating many levels of the stack, we current
have to make special affordances for the /version health check when run
in AWS and we request without a registered domain.  Instead, we set up
an environment that doesn't make any requests to Core for the special
case of /version on AWS. A ticket has been filed to make CurrentDomain
more robust and less pervasive.