CORE-5692 Add and apply feature flag to specify shape file region query limit
Specifies the query limit used for fetching shapefile regions as well as the aggregated that is fetched by the cardVisualizationChoropleth. It is important that the same value is used in both cases to avoid erroneously displaying regions as empty if less aggregated data is fetched than the number of regions.

Add getScalarValue method to ServerConfig with optional default value

Signed-off-by: Randy Antler <randy.antler@socrata.com>