EXPERIMENTAL: Refactor $.toHumaneNumber into old numberFormatter;
This replaces the coffeescript residue in numberFormatter.js (from
janus-govstat) with a modified version of $.toHumaneNumber that
supports optional decimal precision (which was the only reason we
were using numberFormatter). With default options it will always
produce the same output as the old $.toHumaneNumber. Commaification
was also moved into this service. The goal was to reduce the number
of number formatters we use from 2 to 1.0, make use of angular services
instead of jQuery functions, and remove coffeescript because it was
weird and displeasing the linter. Still testing it out and seeing
if it can be cleaned up.