CORE-5841: The Tree map visualization loads blank;
Reasoning:
The setInterval that was initially written only waited for 500ms before
moving forward with the `actualFetchRows`, which appears to be
inadequate for large dataset visualizations. This is all just
conjecture, though. Regardless, the system is now fixed at a 10s
timeout, which seems to be a reasonable amount of waiting time, and in
most cases, the interval will complete light years before that.

QA:
Risk: low
Reproduction:
- Find a large dataset.
- Make a Tree Map, save the new view, and refresh the page.
- You will inconsistently see a failed render of the Tree Map.

Reviewed-by: Stan Rawrysz