CORE-6267: Group by is unapplied for NBE datasets;
Reasoning:
This is another base query issue. When the dataset is viewed by a user
that does not have editing permissions, the base query is set to the
same query that the dataset has. This causes the group by application to
go ignored. My solution is to no longer ignore in this case.

QA:
Risk: low
Reproduction:
- Go to a dataset as a user that does not have editing capabilities.
- Attempt to create a saved view with a group by applied.
- Notice that the group by is no longer applied after viewing the
  completed saved view.