CHART-86 - Intermittent Drag & Drop Test Failure
This ticket's title is actually a misnomer: Drag & Drop was just the first test I observed failing because of this issue.

Here is what I think was causing these intermittent failures:

```
// Hillariously, removing the standard mocks in an `afterEach` appears to have
// introduced a race condition in the tests whereby a `setTimeout` would
// occasionally fire after the call to `removeStandardMocks()` and error out
// because the store it tried to access no longer existed.
//
// Since we are overwriting the same variables on every test run, it seemed
// unnecessary to actually remove them before overwriting them on the next
// test run.
```

(I also updated `removeStandardMocks()` to actually remove all of the stores that we put on the window in case we can figure out a workaround in the future and resume its use).