CORE-6378: Flannel appears upon click with a spinner, then refreshes once row query returns
If any points are clicked, a flannel with a spinner immediately appears and the row query is kicked off.
Once the query responds, flannel content updates:
  - if the query is successful, displays row information in flannel
  - if the query fails, displays an error message.

Flannel can be closed once query resolves (either on successful flannel or error message flannel).
It can not be closed on the spinner flannel while request is pending.

Other changes:
  - flannel code refactoring
  - slightly improved efficiency (execute code only if points are clicked)
  - added tests for featureMapFlannel and CardDataService.getRows()
  - changed getRows() in CardDataService.js to return null if query fails.
  - incorporated load spinner refactoring from Ragan Webber