CORE-6326: Fixed group where issues; Made having behave like where;
Reasoning:
The idea of subtracting queries only works rarely works as expected. In
reality, subtraction should really be uniqueness, but that doesn't
exist. To avoid complicating the implementation, I've approached the
issue with a why-not-both attitude. You can have a base query, and that
will be appended to the queries.

QA:
Risk: High
Reproduction:
- Complicated! See https://socrata.atlassian.net/browse/CORE-6326;