CORE-6214 Fix phidippides making two separate ZK calls for address and port
The Phidippides class now accepts two parameters in the constructor to specify the address and port. If not provided, it checks the ENV for those values, if not found, it asks ZooKeeper for the address/port information in a single call.
Add tests to verify that values are properly sourced from constructor arguments, environment variables, or ZooKeeper

Signed-off-by: Randy Antler <randy.antler@socrata.com>