CORE-5826: NBE Point geometry types successfully use DI2;
- Removed all of the geometry types other than Point, as DI2 doesn't
  handle them yet;
- Added localization for Point validation errors;

Reasoning:
NBE datasets need to have the capability of geocoding, and synthentic
points that are offered by DeltaImporter 2. Unfortunately, at the time
of this commit, DI2 did not support other NBE geometry types. The system
is set up to handle points, but can easily be added to for multipoints,
multipolygons, etc.

QA:
Risk: High
Reproduction:
- Attempt to ingress an NBE dataset through upload.
- Try to geocode a point.
- Try to remap a point to latlong coordinates.
- Try to ingress a column with WKT for points.
- Ensure that there were no OBE ingress regressions.