Feature: SSL by default for OData
Description: With a recent check-in, Core now supports accessing private
datasets through OData.  This means that the customer will be sending
his or her username and password to the API.  If he or she grabs the current
URL, they will get an unsecure, HTTP version of the URL.  This could
potentially compromise their username and password.

This change sets it to HTTPS by default.

Reviewed-By: TBD

QA: Rake tests + Local validation