CORE-6490: Fix dataset column filtering
This regression bug was caused by upgrading from an old version of underscore to a
more modern version of lodash. The meaning of the _ chaining-wrapper function had
changed during the time between these two versions; the functions added to the
chain were no longer being called because the value() method was never invoked,
and the newer versions don't eagerly evaluate the chain without that method call.