CORE-5296: Fix card resize issue;
animateTo was using scope.$watch to watch for changes in the animate-to
attribute, however it wasn't reliably firing during rapid resizes. Turns
out that wrapping the place where cardStates was set in cardLayout in a
safeApply fixed things (probably wasn't triggering a digest if it was
set in the middle of an in-progress digest or whatever angular mumbo
jumbo is involved there).