ENG: Build failing because of config checks
Because we're expecting the CORESERVICE_APP_TOKEN to be enabled
in production mode, docker container builds are failing since
we run `rake assets:precompile` in production mode.

Change it so that we notify airbrake of missing configs that would
cause problems in production. This will not cause unnecessary errors
to be reported in development mode because we don't airbrake in dev
mode.