ONCALL-2790 - Modify calendar autojump to account for more events
Originally, the calendar would use the first requested chunk of events
to calculate the closest calendar event, and then jump to it.  This is
50 events, and likely does not have enough to make a meaningful jump.
The ideal solution would be to make a query for the closest dated event,
but that's beyond the scope of this time-boxed fix.  This solution
allows for a number of requests to come in, accumulating them in an
`_allEvents` array, and then using that after a debounce period to
calculate the jump location.  A debounce is not ideal, as we should
actually be tracking the inflight requests and doing the jump once
they've completed, but again, beyond the time-box scope.  Finally, there
was a logic issue in how the start and end dates were ensured to be Date
objects, so I broke the ternary into if/else statements.