CORE-6588: Fix edge cases with histogram;
Added an isFinite check for svg values in histograms. This was breaking
histograms in the event that the first or last value was undefined. The
chart would try to render NaN as an edge value and the chart would appear
blank. Also fixed a minor issue where if one of the edge values was
close-to-but-not-equal-to zero, it would show up on the axis instead of
at least one pixel away from the axis (like the non-edge values do).