ONCALL-2790 - Remove calendar autojump
The way autojump is engineered relies on dates near the current date to
be in the first block of rows requested.  If this is not the case, the
user will be autojumped to an unexpected date.  If there is customer
pushback from this removal, we will need to add a way to fetch rows
around the current date to determine meaningful autojumping.