Allow bypassing CSRF tokens for specific actions
The following routes no longer validate CSRF tokens for incoming POST request:
 signup XHR via JSON
 login XHR via JSON
 contact dataset owner XHR via JSON

Note that both JSON and DATA format requests are honored due the existing implementation which makes reference to Chrome identifying incoming XHR request as DATA.

It is well understood that disabling CSRF is not a good idea, and this change is intended only as a stop-gap measure to stanch the bleeding for a high-value customer.

The long-term intention is undertake a large-scale refactoring of how session management is implemented after the upcoming summit.

Signed-off-by: Randy Antler <randy.antler@socrata.com>