CORE-6801 - Fix feature map dblclick w/ slow data fetch breaking JS
The underlying issue here was that we were trying to dispose of a
subscription that wasn't guaranteed to exist at the time of disposal,
specifically the `isolateScrollSubscriber`, which would only exist in
the case that the data request for location data completed.  Moved the
disposal to a `$destroy` handler on the scope, which will be disposed of
via the normal means.