CORE-7224 Inject CSRF token in PUT, POST, and PATCH requests
Previously we apparently injected the X-CSRF-Token header, but never bothered to check what _value_ we passed. Fun fact, we weren't passing any value at all.

This change grabs the CSRF token from the socrata-csrf-token cookie and passes it along to Core.

It also passes the requester property along to $http requests

It throws an error when CSRF token is not available in the socrata-csrf-token cookie

Fixes tests broken by throwing on missing CSRF cookie

Signed-off-by: Randy Antler <randy.antler@socrata.com>