ENG: Guard getContentHeight for 'cannot read property getContentHeight';
Reason:
Images rerender the story once loaded to help adjust the story with the
correct image height. As it stands, an image can load before the rich
text editors have completely loaded. This causes a situation where we
rerender even though we don't have rich text editors to query for a
height. In response to this issue, we've opted to add a guard for
getContentHeight, and only run it when we have a RichTextEditor.

Tests:
- None added.