EN-211 Increase button contrast on SDP template list.
The list of templates on the Social Data Player Templates list screen
displays multiple buttons beside each list item. Those buttons were all
greyed out, appearing disabled, until the user hovers over them with the
mouse. As part of the accessibility review, it was identified that when
one of the buttons has tab focus, the focus has insufficient color
contrast to be easily visible.

This change:
- removes the default greyed out appearance from the buttons so they
  allows appear enabled when they are enabled.
- adds an outline to make it easy to distinguish when a button has
  keyboard focus.
- adds the HTML disabled attribute to the Delete button when it is
  disabled so it is not in the tab order.