Increase worker timeout to 60 seconds in dev mode.
This prevents long-running asset compilation from resulting in a 502 when developing.

We should also fix the situation where the `th-breakpoint` gem cannot be installed and causes the sass compiler to whine; this PR does not address this issue, which is one of the ways that we end up manifesting the timeout.