EN-1232 Update focus highlighting for admin.
Changes the style that is applied to some HTML elements when they have
focus. This is intended to address accessibility concerns by making it
easier to see which controls have keyboard forcus when a user is tabbing
through controls on the page.

The style update only applies to button like elements.