BUG: Globally strip byte order marks from merged stylesheets.
When we were concatenating the contents of multiple style files, Ruby was deciding to insert a separate BOM at the beginning of each file, which in the merged styles ended up causing the contents of every constituent stylesheet beyond the first to be ignored by the browser. This fix strips all BOMs from rendered & merged stylesheets.