ENG: Add test coverage for BOM stripping
The version of Rails we're using seems not to support rendering response bodies in functional tests unless a template is being rendered.

In the StylesController, we render the text directly, without a template, so I pulled out the code that builds the merged stylesheets into a protected helper method to make it testable.

Less than ideal, but I really want test coverage for this issue. Open to other ideas if anyone has any suggestions. That is, other than writing integration tests. ;-)

Signed-off-by: Randy Antler <randy.antler@socrata.com>