BUG: Rails converting empty arrays in json into nils
According to https://github.com/rails/rails/pull/8862, Rails will automatically reach deep into any JSON in `params` and convert instances of `[]` into `nil`. This is to cover up a bug in `ActiveRecord` that is perceived as a security issue.

This means, however, that we are not persisting what is actually sent by the client, and in the case of VIF it means that we are breaking the spec for (at least) the `filters` property, which must be an array and is empty if there are no filters. Rather than change the VIF spec to allow for `null` OR `[]`, we instead re-assign `filters` to an empty array if Rails has gone behind our back and converted it into `nil` even though this is not what the client actually sent. WTF Rails.

- Ensure that when we read or write a block, all components that are VIFs have an empty array to indicate the absence of filters.
- Add test.
- Update `README.md` to include instructions for running one spec file only.