CORE-5190: Fix feature map with one point;
The reason the render would fail is the backend would return an
undefined extent.  CardDataService would refuse to parse this and
throw an error, and the feature map would spin forever before timing
out on render.

This has been changed so that CardDataService will return undefined
if the extent is an empty object.  FeatureMapController has also
been changed to use the defaultExtent if the serverExtent is
undefined (which I believe was a bug).

QA: Ensure that feature maps that have only one point work.