EN-1462: Bugfix for cards loading newer schema versions;
Cards inherit their version from their page.  Pages don't validate
their schemas but cards do.  Currently, if no schema exists for the
card's version an error is thrown.  This changes things so that if
the card's version is newer than the latest schema version, it will
use that instead.  This is meant to alleviate the situation where
a new schema version is added, the page with the new version is saved,
and then the code (and by extension, the new schema) is rolled back.
Rather than exploding with an error, the card will now attempt to
coerce the new metadata to the old metadata.