EN-97: Add pagination to grid footer. NBE-only.
Because performance when scrolling to an arbitrary position in a large
dataset is poor, even in NBE, we use pagination to prevent this. When
navigating to the last page, we use a reverse sort (default is now :id)
in order to performantly handle this case.

This is governed by the feature flag `nbe_bucket_size`, which unless set
to `false`, will always be set to a number. If set to false, then we
will not use buckets and will not be paging. `nbe_bucket_size` was not
originally purposed for this, but I've linked them because it makes
sense to me.

It is expected that, when clicking Last Page, you will scroll to the
BOTTOM of the page. This is intended behavior. Similarly, when otherwise
paging, you will be taken to the TOP of the page.

Risk: Medium. I feel that it's been tested fairly thoroughly, but since
no tests cover this code, it's hard to be confident.
Regression: Should make sure anything that uses the core Frontend Old UX
data-loading code (visualizations, DataSlate) continues to work.
Test Cases: See comments in EN-97.