EN-2100: Fix feature map triggering unsaved state;
The serverExtent observable shouldn't start with undefined, because
this triggers the extent to fire twice, once with the whole world,
then with the correct extent.  This is interpreted as an extent
change.  Instead, it makes more sense to only produce a value from
the serverExtent observable when the request completes.  See EN-1511.