EN-2010 - Add userzoom for admin and profile pages
This adds a third-party modal dialog to `/admin/*` and `/profile/*` pages
asking the user if they would like to participate in a survey around the admin
profile editing experience.

The survey code on the profile page is only enabled for members of the
domain, to avoid [Paperwork Reduction
Act](https://en.wikipedia.org/wiki/Paperwork_Reduction_Act) complications.

Dismissing the modal sets two cookies that start with UZ.

** Possible impact to automated testing. **