EN-2154: Add endpoint for consul check
This consul check will determine if the currently running
rails app is serving the version of the storyteller app
that matches what is set in the consul k/v store. This
allows us to have multiple versions of the app running
in the same environment and only route traffic to all
instances running one specific version.