EN-3189 Make RTE Toolbar Accessible to Some IE11 Users
Some versions of IE11 do not appear to support overriding `pointer-events: none` on child elements; this PR removes `pointer-events: none;` on the root RTE wrapper to allow users of earlier versions of IE to actually click things.

This PR (also, unrelatedly) adds a closing `</ul>` tag to the edit template, which was causing IE to complain in the console.