EN-2469: Fix Choropleth after frontend-viz version bump
frontend-visualizations was updated by EN-2491 to require additional
properties in the VIF, as well as requiring us to send the VIF along
with every call to render. This commit updates the Data Lens choropleth
directive so it can continue to work, but we'll need to finish updating
it to take all of the changes into account.