'Create Story' button does not work in classic browse.
Clicking the `create story` button in the classic browse mode (not browse2) does not cause a story to be created.

A previous change to fix exceptions in IE moved two function declarations up one scope level; this move inadvertently did not include the final closing brace, leading to the Ajax call never being fired as it was suddenly inside the success callback for the Ajax call.