EN-2552: Upgrade socrata-visualizations;
This upgrades socrata-visualizations to 7.0.0. The breaking change
in 7.0.0 is that leaflet is no longer included in the bundle and
must be supplied by the host application. In this case, frontend
is including leaflet on the page using the bower-all jammit bundle,
so we also declare leaflet as external in the frontend webpack config
("L": "L" in package.json 'dataLensWebpackExternals' key).

After this change, the Data Lens webpack bundle should be smaller as
we will no longer be including 2 versions of leaflet on Data Lens.

QA: Make sure that maps still work on Data Lens pages.
(cherry picked from commit bf3ea26c29a38156d1262aa16c09bd261cd65f27)