EN-3640 Cannot choose datasets from federated domains in Safari, IE11 or FF
Basically, when we asked for `/api/migrations/<four-four>.json` on the host domain, we would get a 302 back from the server pointing to the correct domain. For whatever reason, Chrome was happy to follow this automatically but the typical AJAX behavior is not to follow 302s resulting from AJAX regardless of CORS settings on the redirect target domain.

This change fetches both the domain and the four-by-four for the selected dataset and then crafts a request directly to the federated domain instead. Because of the CORS settings on the `/api/migrations/*` and `/api/views/*` endpoints, things should work correctly.