EN-3926 - Link to story stat page from /admin/analytics is wrong.
The link from `/admin/analytics` points to `/stories/s/<four-four>/stats`, which is actually incorrect. We actually want to send people to `/d/<four-four>/stats`. Redirecting from the endpoint in `/stories/s/` allows us flexibility in the future in case we want to have a more stories-specific format at the cost of slightly longer page load times (on account of being redirected) for the time being.

- Add route for `/stories/s/<four-four>/stats`.
- Add `stories_controller` method to redirect users to `/d/<four-four>/stats`.
- Redirect unauthenticated users to login.
- Add tests.