EN-1853: Warning when no choropleth polygons to render;
QA: Load a choropleth for which there are no polygons intersecting the extent of the data.
Instead of seeing a blank map, you should see a very helpful error message.

Note that if there are no points inside the polygons but they DO intersect the extent of the data,
polygons will render as empty and this warning will not be displayed.