EN-4829: Workers don't work when not active version
Storyteller has the concept of active versions when deployed so
that we can pre-deploy an app and then switch routing so that all
traffic is simultaneously sent to the new instance. Since moving
the queue processing into the docker container with runit, we
ran into the situation where we would have inactive versions of
storyteller processing items off the job queue. This change prevents
that situation.