EN-327: data_lens_choropleth_custom_boundary;
This is a new feature flag that overrides the polygon that the Data Lens choropleth uses to fetch
regions.  The current behavior is to use the extent of the points to get a rectangle that is a rough
estimate of where interesting regions will be.  However, there are lots of problems here.  Sometimes
we fetch way too many regions, sometimes we leave out regions, and sometimes this can have interesting
consequences with our 5k region limit.  Allowing CS to specify a custom polygon in WKT format in the
internal panel will allow fine-grained control over which regions are fetched.  It also eliminates an
HTTP request when fetching the data for the choropleth, which improves render times.

QA: Set the data_lens_choropleth_custom_boundary feature flag to a polygon in the WKT format.  Use
http://arthur-e.github.io/Wicket/sandbox-gmaps3.html to generate such a string.  Then ensure that
Data Lens choropleths only render regions inside this polygon.  If the feature flag is disabled, the
old method should be used.